scripts/seo_generator/pyproject.toml
[project]
name = "seo-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "openai>=1.84.0",
    "openpyxl>=3.1.5",
    "python-frontmatter>=1.1.0",
    "selenium>=4.33.0",
]
